Remove demographics file
All information and more is present in the the Quality Report. We will
choose to use only one dataset instead of two.

Note: The demographics file contained dates and such, while the Quality
Report doesn't. We will use the later by now and, if the results come out
worse, we revert to the demographics.